// src/models/login/memauthRequest.model.ts
export class MemauthRequest {
  useridin?: string;
  firstname?: string;
  lastname?: string;
  DOB?: string;
  email?: string;
  mobile?: string;
  hintAnswer?: string;
  hintQuestion?: string;
  phoneNumberType?: string;
  // below 2 fields are required in memeber info page.
  memberid?: string;
  suffix?: string;
  username?: string;
}